plugins/agents/agent-toolkit/pyproject.toml
[project]
name = "hopeit-agents.agent-toolkit"
version = "0.1.0"

description = "Hopeit Agents: Agent Toolkit plugin"
dynamic = ["readme"]

dependencies = [
    "hopeit.engine>=0.27.0",
    "hopeit-agents.mcp-client",
]

license = { text = "Apache 2" }
authors = [
    { name = "hopeit-agents contributors", email = "opensource@hopeit.com.ar" }
]
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Development Status :: 3 - Alpha",
    "Framework :: AsyncIO",
]

[project.urls]
"Homepage" = "https://github.com/hopeit-git/hopeit.agents"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
readme = { file = ["README.md"], content-type = "text/markdown" }
